fix: тесты 3

+++ controllers/cards.js
@@ -15,23 +15,50 @@ const getCards = (req, res) => {
 };
 
 const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then(() => res.send({ message: 'Карточка удалена' }))
-    .catch(() => res.status(400).send({ message: 'На сервере произошла ошибка' }));
+  if (req.params.cardId.length === 24) {
+    Card.findByIdAndRemove(req.params.cardId)
+      .then((card) => {
+        if (card) {
+          res.send({ message: 'Карточка удалена' });
+        } else {
+          res.status(404).send({ message: 'Карточка с таким Id не найдена' });
+        }
+      });
+  } else {
+    res.status(400).send({ message: 'Некоректный Id карточки' });
+  }
 };
 
 const likeCard = (req, res) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .populate(['owner', 'likes'])
-    .then((card) => res.send(card))
-    .catch(() => res.status(400).send({ message: 'На сервере произошла ошибка' }));
+  if (req.params.cardId.length === 24) {
+    Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
+      .populate(['owner', 'likes'])
+      .then((card) => {
+        if (card) {
+          res.send(card);
+        } else {
+          res.status(404).send({ message: 'Карточка с таким Id не найдена' });
+        }
+      });
+  } else {
+    res.status(400).send({ message: 'Некоректный Id карточки' });
+  }
 };
 
 const dislikeCard = (req, res) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .populate(['owner', 'likes'])
-    .then((card) => res.send(card))
-    .catch(() => res.status(400).send({ message: 'На сервере произошла ошибка' }));
+  if (req.params.cardId.length === 24) {
+    Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
+      .populate(['owner', 'likes'])
+      .then((card) => {
+        if (card) {
+          res.send(card);
+        } else {
+          res.status(404).send({ message: 'Карточка с таким Id не найдена' });
+        }
+      });
+  } else {
+    res.status(400).send({ message: 'Некоректный Id карточки' });
+  }
 };
 
 module.exports = {
